pyproject.toml
[tool.poetry]
name = "mysql"
version = "0.1.0"
description = ""
authors = ["Sergio Semprebom <sergiosemprebom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.1.0"
mysqlclient = "^2.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
